actions: Split out zerver.actions.hotspots.
(cherry picked from commit 12130da339941f49264b4b2fbed7689d2f54005c)
